added test for root privileges

+++ contrib/update-full.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
-#
-# run this as root!
-#
+if [ $(id -u) -ne 0 ] ; then
+	echo "This script must be run as root.";
+	exit 1
+fi
 apt-get update; apt-get -y dist-upgrade; apt-get autoremove; apt-get autoclean
